Cloud Functions/fetch-data/requirements.txt
google-cloud
google-cloud-storage
kaggle
